add Ioc support

+++ src/com/qa/framework/TestCaseBase.java
@@ -1,18 +1,18 @@
 package com.qa.framework;
 
-import com.qa.framework.config.DriverConfig;
-import com.qa.framework.config.PropConfig;
 import com.qa.framework.cache.DriverCache;
 import com.qa.framework.cache.MethodCache;
+import com.qa.framework.config.DriverConfig;
+import com.qa.framework.config.PropConfig;
 import com.qa.framework.data.SuiteData;
+import com.qa.framework.ioc.ClassFinder;
+import com.qa.framework.library.base.StringHelper;
 import com.qa.framework.testngListener.PowerEmailableReporter;
 import com.qa.framework.testngListener.TestResultListener;
-import com.qa.framework.library.base.StringHelper;
 import org.apache.log4j.Logger;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.remote.DesiredCapabilities;
 import org.openqa.selenium.remote.RemoteWebDriver;
-import org.reflections.Reflections;
 import org.testng.annotations.*;
 
 import java.lang.reflect.Constructor;
@@ -20,7 +20,7 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.Set;
+import java.util.List;
 
 import static com.qa.framework.ioc.AutoInjectHelper.initFields;
 
@@ -31,7 +31,6 @@ import static com.qa.framework.ioc.AutoInjectHelper.initFields;
 public abstract class TestCaseBase {
     public WebDriver driver;
     protected Logger logger = Logger.getLogger(this.getClass());
-    protected Reflections reflections = new Reflections("com.qa");
     private SuiteData suiteData = null;
 
     public WebDriver getDriver() {
@@ -42,9 +41,9 @@ public abstract class TestCaseBase {
     public void BeforeSuite() throws Exception {
         logger.info("beforeSuite");
         HelperLoader.init();
-        Set<Class<? extends SuiteData>> subTypes = reflections.getSubTypesOf(SuiteData.class);
-        if (subTypes.size() == 1) {
-            Class<? extends SuiteData> clazz = subTypes.iterator().next();
+        List<Class<?>> subTypes = ClassFinder.getClassListBySuper(SuiteData.class);
+        if (subTypes.size() > 0) {
+            Class<?> clazz = subTypes.get(0);
             Constructor<?> con = clazz.getConstructor();
             if (con != null) {
                 suiteData = (SuiteData) con.newInstance();
@@ -67,9 +66,9 @@ public abstract class TestCaseBase {
         if (PropConfig.getCoreType().equalsIgnoreCase("ANDROIDAPP") || PropConfig.getCoreType().equalsIgnoreCase("IOSAPP")) {
             driver.quit();
         }
-        Set<Class<? extends SuiteData>> subTypes = reflections.getSubTypesOf(SuiteData.class);
-        if (subTypes.size() == 1) {
-            Class<? extends SuiteData> clazz = subTypes.iterator().next();
+        List<Class<?>> subTypes = ClassFinder.getClassListBySuper(SuiteData.class);
+        if (subTypes.size() > 0) {
+            Class<?> clazz = subTypes.get(0);
             Constructor<?> con = clazz.getConstructor();
             if (con != null) {
                 suiteData = (SuiteData) con.newInstance();
@@ -86,23 +85,25 @@ public abstract class TestCaseBase {
     @org.testng.annotations.BeforeClass(alwaysRun = true)
     public void BeforeClass(@Optional String browser, @Optional String hubURL) throws Exception {
         logger.info("beforeClass");
-        if (!(PropConfig.getCoreType().equalsIgnoreCase("ANDROIDAPP") || PropConfig.getCoreType().equalsIgnoreCase("IOSAPP"))) {
-            if (hubURL != null) {
-                DesiredCapabilities capability = null;
-                if (browser.contains("firefox")) {
-                    capability = DesiredCapabilities.firefox();
-                } else if (browser.contains("chrome")) {
-                    capability = DesiredCapabilities.chrome();
+        if(!isUnitTest()){
+            if (!(PropConfig.getCoreType().equalsIgnoreCase("ANDROIDAPP") || PropConfig.getCoreType().equalsIgnoreCase("IOSAPP"))) {
+                if (hubURL != null) {
+                    DesiredCapabilities capability = null;
+                    if (browser.contains("firefox")) {
+                        capability = DesiredCapabilities.firefox();
+                    } else if (browser.contains("chrome")) {
+                        capability = DesiredCapabilities.chrome();
+                    }
+                    try {
+                        driver = new RemoteWebDriver(new URL(hubURL), capability);
+                    } catch (MalformedURLException e) {
+                        e.printStackTrace();
+                    }
+                } else {
+                    driver = DriverConfig.getDriverObject();
                 }
-                try {
-                    driver = new RemoteWebDriver(new URL(hubURL), capability);
-                } catch (MalformedURLException e) {
-                    e.printStackTrace();
-                }
-            } else {
-                driver = DriverConfig.getDriverObject();
+                DriverCache.set(driver);
             }
-            DriverCache.set(driver);
         }
         initFields(this);
         beforeClass();
@@ -111,11 +112,17 @@ public abstract class TestCaseBase {
     public void beforeClass() {
     }
 
+    public boolean isUnitTest() {
+        return false;
+    }
+
     @AfterClass(alwaysRun = true)
     public void AfterClass() {
         logger.info("afterClass");
-        if (!(PropConfig.getCoreType().equalsIgnoreCase("ANDROIDAPP") || PropConfig.getCoreType().equalsIgnoreCase("IOSAPP"))) {
-            driver.quit();
+        if(!isUnitTest()) {
+            if (!(PropConfig.getCoreType().equalsIgnoreCase("ANDROIDAPP") || PropConfig.getCoreType().equalsIgnoreCase("IOSAPP"))) {
+                driver.quit();
+            }
         }
         afterClass();
     }
@@ -132,8 +139,10 @@ public abstract class TestCaseBase {
             currentMethodName = method.getName();
         }
         MethodCache.set(StringHelper.removeSpecialChar(currentMethodName));
-        if (!(PropConfig.getCoreType().equalsIgnoreCase("ANDROIDAPP") || PropConfig.getCoreType().equalsIgnoreCase("IOSAPP"))) {
-            driver.manage().deleteAllCookies();
+        if(!isUnitTest()) {
+            if (!(PropConfig.getCoreType().equalsIgnoreCase("ANDROIDAPP") || PropConfig.getCoreType().equalsIgnoreCase("IOSAPP"))) {
+                driver.manage().deleteAllCookies();
+            }
         }
         beforeMethod(method, para);
     }
